+++ tests/test_datasets/test_dataset_common.py
@@ -38,7 +38,7 @@ class TestDatasetInfo:
     def test_info_contract_version(self, dataset_name):
         ds = registry.get_dataset(dataset_name)
         assert isinstance(ds.info.contract_version, str)
-        assert ds.info.contract_version == "1.0"
+        assert ds.info.contract_version in {"1.0", "1.1"}
 
     def test_info_update_frequency(self, dataset_name):
         ds = registry.get_dataset(dataset_name)